# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API user management Server by Fiber | Doc by Swagger.",
        "title": "Talardnad API",
        "contact": {
            "name": "admin",
            "url": "http://subalgo.com/support",
            "email": "admin@subalgo.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
...
        "/slots/delete/{id}": {
            "delete": {
                "description": "Delete slot",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "slots"
                ],
                "summary": "Delete slot",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Slot ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/slots/delete/{id}/zone/{zoneID}/date/{date}": {
            "delete": {
                "description": "Delete slot by date and zone",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "slots"
                ],
                "summary": "Delete slot by date and zone",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Slot ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Zone ID",
                        "name": "zoneID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Date",
                        "name": "date",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/slots/edit/{id}": {
            "patch": {
                "description": "Edit slot",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "slots"
                ],
                "summary": "Edit slot",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Slot ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Slot update data",
                        "name": "updateDTO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.SlotUpdateDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.Slot"
                        }
                    }
                }
            }
        },
        "/slots/get/{id}": {
            "get": {
                "description": "Get all slots",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "slots"
                ],
                "summary": "Get all slots",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Market ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.Slot"
                        }
                    }
                }
            }
        },
        "/slots/markets/{marketID}/date/{date}": {
            "get": {
                "description": "Get slots by date",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "slots"
                ],
                "summary": "Get slots by date",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MarketID",
                        "name": "marketID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Date",
                        "name": "date",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.Slot"
                            }
                        }
                    }
                }
            }
        },
        "/slots/provider/get/{id}": {
            "get": {
                "description": "Get provider slots",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "slots"
                ],
                "summary": "Get provider slots",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Provider ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.Slot"
                            }
                        }
                    }
                }
            }
        },
        "/slots/{marketId}/create": {
            "post": {
                "description": "Create or update layout",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "slots"
                ],
                "summary": "Create or update layout",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Market ID",
                        "name": "marketId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Layout data",
                        "name": "layout",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.LayoutRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dtos.LayoutRequest": {
            "type": "object",
            "properties": {
                "layout": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.ZoneLayout"
                    }
                }
            }
        },
...
        "dtos.SlotUpdateDTO": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/entities.Category"
                },
                "height": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "status": {
                    "$ref": "#/definitions/entities.SlotStatus"
                },
                "width": {
                    "type": "number"
                }
            }
        },
        "dtos.Stall": {
            "type": "object",
            "properties": {
                "height": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "stallType": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                }
            }
        },
        "dtos.ZoneLayout": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "stalls": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.Stall"
                    }
                },
                "zone": {
                    "type": "string"
                }
            }
        },
        "entities.Booking": {
            "type": "object",
            "properties": {
                "booking_date": {
                    "description": "Changed from Date to BookingDate",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "market_id": {
                    "type": "string"
                },
                "method": {
                    "$ref": "#/definitions/entities.Method"
                },
                "payment": {
                    "$ref": "#/definitions/entities.Payment"
                },
                "price": {
                    "type": "number"
                },
                "slot": {
                    "$ref": "#/definitions/entities.Slot"
                },
                "slot_id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/entities.BookingStatus"
                },
                "updated_at": {
                    "type": "string"
                },
                "vendor": {
                    "$ref": "#/definitions/entities.Vendor"
                },
                "vendor_id": {
                    "type": "string"
                }
            }
        },
        "entities.BookingStatus": {
            "type": "string",
            "enum": [
                "pending",
                "cancelled",
                "completed"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusCancelled",
                "StatusCompleted"
            ]
        },
        "entities.Category": {
            "type": "string",
            "enum": [
                "clothes",
                "food",
                "crafts",
                "produce",
                "electronics",
                "services",
                "other"
            ],
            "x-enum-varnames": [
                "CategoryClothes",
                "CategoryFood",
                "CategoryCrafts",
                "CategoryProduce",
                "CategoryElectronics",
                "CategoryServices",
                "CategoryOther"
            ]
        },
...
        "entities.Method": {
            "type": "string",
            "enum": [
                "PromptPay"
            ],
            "x-enum-varnames": [
                "MethodPromptPay"
            ]
        },
        "entities.Payment": {
            "type": "object",
            "properties": {
                "booking": {
                    "$ref": "#/definitions/entities.Booking"
                },
                "booking_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "method": {
                    "$ref": "#/definitions/entities.Method"
                },
                "payment_date": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "status": {
                    "$ref": "#/definitions/entities.PaymentStatus"
                },
                "transactions": {
                    "$ref": "#/definitions/entities.Transaction"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "entities.PaymentStatus": {
            "type": "string",
            "enum": [
                "pending",
                "completed",
                "failed"
            ],
            "x-enum-varnames": [
                "PaymentPending",
                "PaymentCompleted",
                "PaymentFailed"
            ]
        },
        "entities.Slot": {
            "type": "object",
            "required": [
                "date"
            ],
            "properties": {
                "booking": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Booking"
                    }
                },
                "category": {
                    "$ref": "#/definitions/entities.Category"
                },
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "market_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "status": {
                    "$ref": "#/definitions/entities.SlotStatus"
                },
                "updated_at": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                },
                "zone": {
                    "type": "string"
                }
            }
        },
        "entities.SlotStatus": {
            "type": "string",
            "enum": [
                "available",
                "booked",
                "maintenance"
            ],
            "x-enum-varnames": [
                "StatusAvailable",
                "StatusBooked",
                "StatusMaintenance"
            ]
        },
        "entities.Transaction": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "method": {
                    "type": "string"
                },
                "payment": {
                    "$ref": "#/definitions/entities.Payment"
                },
                "payment_id": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "ref1": {
                    "type": "string"
                },
                "ref2": {
                    "type": "string"
                },
                "ref3": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/entities.TransactionStatus"
                },
                "transaction_date": {
                    "type": "string"
                },
                "transaction_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "entities.TransactionStatus": {
            "type": "string",
            "enum": [
                "pending",
                "completed",
                "failed"
            ],
            "x-enum-varnames": [
                "TransactionPending",
                "TransactionCompleted",
                "TransactionFailed"
            ]
        },
        "entities.Vendor": {
            "type": "object",
            "properties": {
                "bookings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Booking"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}